+++ lib/prima_auth0_ex/jwks_strategy.ex
@@ -5,13 +5,11 @@ defmodule PrimaAuth0Ex.JwksStrategy do
 
   alias PrimaAuth0Ex.Config
 
+  alias PrimaAuth0Ex.OpenIDConfiguration
   use JokenJwks.DefaultStrategyTemplate
 
   def init_opts(opts) do
-    Keyword.merge(opts, jwks_url: jwks_url())
+    jwks_url = Config.server(:auth0_base_url) |> OpenIDConfiguration.fetch!() |> Map.fetch!(:jwks_uri)
+    Keyword.merge(opts, jwks_url: jwks_url)
   end
-
-  defp jwks_url, do: base_url() <> "/.well-known/jwks.json"
-
-  defp base_url, do: Config.server!(:auth0_base_url)
 end